Update pongServer.py
Added some basic code to check sync and send the appropriate gamestate to both clients.

Added a function to swap paddle1 coordinates and paddle2 coordinates in the gamestate, largely for testing purposes. Hopefully this can be removed later.

The intent of the swap function is to address an issue I foresee with the way clients send the gamestate. The client responsible for the right side paddle sends the coordinates backwards compared to the standardized format, based on how the client sends the gamestate.